Fixed: number consts could not be specified to be int64
Changed notation for constant numbers.  Octal numbers are notated using a 'q', bytes (octets) using a 't'.

+++ bootstrap/apt.cpp
@@ -600,10 +600,9 @@ BaseNumberNode::isImaginary() const
 
 //----------------------------------------------------------------------------
 
-IntNode::IntNode(const SrcPos& srcpos, int value,
-                 bool isImaginary,
+IntNode::IntNode(const SrcPos& srcpos, int64_t value, bool isImaginary,
                  const Type& type)
-  : NumberNode<int>(srcpos, value, isImaginary, type)
+  : NumberNode<int64_t>(srcpos, value, isImaginary, type)
 {
 }
